(BSR)[PRO] refactor: rename CollectiveOfferSummaryCreation screen file to fit its exported component naming
